Demo app tests: move stype helpers tests to separate files
