+++ src/shared/helpers/style/__tests__/rule.test.ts
@@ -1,0 +1,8 @@
+import { rule } from '../';
+
+describe('(shared/helpers/style) rule', () => {
+  it('should not change props', () => {
+    const props = {};
+    expect(rule(props)).toBe(props);
+  });
+});
